+++ packages/atmosphere/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.13.2] - 2025-07-14
+
+### Fixed
+
+- Fixed `PrecomputedTexturesGenerator` incorrectly reads FP32 values from half-float textures on the CPU.
+
 ## [0.13.1] - 2025-07-14
 
 ### Fixed